Move base16 and base32 specific bits to their own file.
